third_party/blink/web_tests/external/wpt/css/fill-stroke/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-decoration-fill has initial value match-text
  assert_true: text-decoration-fill doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-fill does not inherit
  assert_true: expected true got false
[FAIL] Property text-decoration-stroke has initial value match-text
  assert_true: text-decoration-stroke doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-stroke does not inherit
  assert_true: expected true got false
Harness: the test ran to completion.
